% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_count.R
\name{sent_counts}
\alias{sent_counts}
\title{Get a count of the number of articles for a specified media source.}
\usage{
sent_counts(source, topic, freq)
}
\arguments{
\item{source}{The news source for which data is sought. Data can either be drawn from all sources (sourc=all), from News24 only (source=news24), or from Business Day only (source=businessday)..}

\item{topic}{The topic which you are interested in.}

\item{freq}{The frequency on which to aggregate the index. Available values: day, week, month.}

\item{model}{The name of the model to use. Even the general index (i.e. no topics) requires a model name.}
}
\value{
tibble.
}
\description{
Get a count of the number of articles for a specified media source.
}
\examples{
# Get a count of the number of articles for a specified media source, model, topic, and frequency of aggregation.
\dontrun{
sent_counts(
   source="all",
   model="model_2021-05-15",
   topic="global",
   freq="month",
   )
}

}
