% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_date_polarity.R
\name{sent_date_polarity}
\alias{sent_date_polarity}
\title{Get a count of the number of positive and negative articles.}
\usage{
sent_date_polarity(source, topic, freq, dict, aggr)
}
\arguments{
\item{source}{The news source for which data is sought. Data can either be drawn from all sources (sourc=all), from News24 only (source=news24), or from Business Day only (source=businessday)..}

\item{topic}{The topic which you are interested in.}

\item{freq}{The frequency on which to aggregate the index. Available values: day, week, month.}

\item{dict}{The dictionary to use in sentiment characterization for the index. Available values: afinn, anew, geninqposneg, huliu, loughran, lsd2015, nrc, vader.}

\item{aggr}{The aggregation method to use. Available values: sent_abspropdiff, sent_relpropdiff, sent_logit, sent_vader, sent_valence.}

\item{model}{The name of the model to use. Even the general index (i.e. no topics) requires a model name.}
}
\value{
tibble.
}
\description{
Get a count of the number of positive and negative articles.
}
\examples{
# Get a count of the number of positive and negative articles for a specified model, topic, dictionary, aggregation method, and frequency.
\dontrun{
sent_date_polarity(
   source="all",
   model="model_2021-05-15",
   topic="global",
   freq="month",
   dict="loughran",
   aggr="sent_logit"
   )
}

}
