% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_index.R
\name{sent_index}
\alias{sent_index}
\title{Pull a specific media sentiment index}
\usage{
sent_index(source, topic, freq, dict, aggr)
}
\arguments{
\item{source}{The news source for which data is sought. Data can either be drawn from all sources (sourc=all), from News24 only (source=news24), or from Business Day only (source=businessday)..}

\item{topic}{The topic which you are interested in.}

\item{freq}{The frequency on which to aggregate the index. Available values: day, week, month.}

\item{dict}{The dictionary to use in sentiment characterization for the index. Available values: afinn, anew, geninqposneg, huliu, loughran, lsd2015, nrc, vader.}

\item{aggr}{The within article aggregation method to use. Available values: sent_abspropdiff, sent_relpropdiff, sent_logit, sent_vader, sent_valence.}

\item{model}{The name of the model to use. Even the general index (i.e. no topics) requires a model name.}
}
\value{
tibble.
}
\description{
Pull a specific media sentiment index
}
\examples{
# Get global sentiment index.
\dontrun{
sent_index(
   source="all",
   model="model_2021-05-15",
   topic="global",
   freq="month",
   dict="loughran",
   aggr="sent_logit"
   )
}

}
